backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-telegram-bot==20.8
twilio==8.11.0
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
